Move test to spec-helpers for clarity and rename examples to __examples__
