+++ lib/src/linting/rules/has-request-payload.spec.ts
@@ -7,7 +7,7 @@ import {
   RequestNode
 } from "../../models/nodes";
 import { TypeKind } from "../../models/types";
-import { fakeLocatable } from "../../test/fake-locatable";
+import { fakeLocatable } from "../../spec-helpers/fake-locatable";
 import { hasRequestPayload } from "./has-request-payload";
 
 describe("rule: has-request-payload", () => {
