+++ lib/src/linting/rules/has-response-payload.spec.ts
@@ -7,7 +7,7 @@ import {
   ResponseNode
 } from "../../models/nodes";
 import { TypeKind } from "../../models/types";
-import { fakeLocatable } from "../../test/fake-locatable";
+import { fakeLocatable } from "../../spec-helpers/fake-locatable";
 import { hasResponsePayload } from "./has-response-payload";
 
 describe("rule: has-response-payload", () => {
