+++ lib/src/parsers/__examples__/contract-endpoint.ts
