+++ lib/src/parsers/__examples__/contract.ts
