+++ lib/src/parsers/__examples__/recursive-imports-and-exports/contract.ts
