+++ lib/src/parsers/__examples__/recursive-imports-and-exports/exported-endpoint-2.ts
