+++ lib/src/parsers/__examples__/recursive-imports-and-exports/imported-endpoint-1.ts
