+++ lib/src/parsers/__examples__/recursive-imports-and-exports/imported-endpoint-2.ts
