+++ lib/src/parsers/nodes/default-response-parser.spec.ts
@@ -1,5 +1,5 @@
 import { ClassDeclaration } from "ts-morph";
-import { createSourceFile } from "../../test/helper";
+import { createSourceFile } from "../../spec-helpers/helper";
 import { parseDefaultResponse } from "./default-response-parser";
 
 describe("@defaultResponse parser", () => {
