+++ lib/src/spec-helpers/fake-locatable.ts
