+++ lib/src/verifiers/nodes/endpoint-verifier.spec.ts
@@ -1,7 +1,7 @@
 import { HttpMethod } from "../../models/http";
 import { BodyNode, EndpointNode, PathParamNode } from "../../models/nodes";
 import { TypeKind } from "../../models/types";
-import { fakeLocatable } from "../../test/fake-locatable";
+import { fakeLocatable } from "../../spec-helpers/fake-locatable";
 import { verifyEndpointNode } from "./endpoint-verifier";
 
 describe("endpoint node verifier", () => {
