+++ lib/src/verifiers/nodes/header-verifier.spec.ts
@@ -1,6 +1,6 @@
 import { HeaderNode } from "../../models/nodes";
 import { TypeKind } from "../../models/types";
-import { fakeLocatable } from "../../test/fake-locatable";
+import { fakeLocatable } from "../../spec-helpers/fake-locatable";
 import { verifyHeaderNode } from "./header-verifier";
 
 describe("header node verifier", () => {
