+++ lib/src/verifiers/nodes/query-param-verifier.spec.ts
@@ -1,6 +1,6 @@
 import { QueryParamNode } from "../../models/nodes";
 import { TypeKind } from "../../models/types";
-import { fakeLocatable } from "../../test/fake-locatable";
+import { fakeLocatable } from "../../spec-helpers/fake-locatable";
 import { verifyQueryParamNode } from "./query-param-verifier";
 
 describe("query param node verifier", () => {
