+++ lib/src/verifiers/nodes/test-verifier.spec.ts
@@ -7,7 +7,7 @@ import {
   TestRequestNode
 } from "../../models/nodes";
 import { TypeKind } from "../../models/types";
-import { fakeLocatable } from "../../test/fake-locatable";
+import { fakeLocatable } from "../../spec-helpers/fake-locatable";
 import { verifyTestNode } from "./test-verifier";
 
 describe("test node verifier", () => {
